requirements.txt
# Dependências principais da aplicação
Flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
pydantic>=2.5.0
python-multipart==0.0.6

# Dependências de segurança e validação
bleach==6.1.0
cryptography==41.0.7

# Dependências de monitoramento
prometheus-client==0.19.0

# Dependências de backup
schedule==1.2.0

# Dependências de desenvolvimento e teste
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0